pyproject.toml
[tool.poetry]
name = "slope-rag-agent"
version = "0.1.0"
description = "RAG agent for slope stability engineering"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
numpy = "^1.26.0"
pandas = "^2.2.0"
scikit-learn = "^1.4.0"
faiss-cpu = "^1.7.4"
rank-bm25 = "^0.2.2"
sentence-transformers = "^2.3.1"
torch = "^2.2.0"
transformers = "^4.37.0"
accelerate = "^0.26.0"
bitsandbytes = "^0.42.0"
pdfplumber = "^0.10.3"
unstructured = "^0.12.4"
jieba = "^0.42.1"
openai = "^1.12.0"
tiktoken = "^0.6.0"
elasticsearch = "^8.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
